% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDE.R
\name{simulateDE}
\alias{simulateDE}
\alias{simulateDE,ANY-method}
\alias{simulateDE,SummarizedExperiment-method}
\alias{simulateDE,SingleCellExperiment-method}
\title{Simulate DE by swapping features}
\usage{
simulateDE(x, which_cols, ...)

\S4method{simulateDE}{ANY}(x, which_cols, prop_DE = 0.01)

\S4method{simulateDE}{SummarizedExperiment}(x, which_cols, ..., use_assay = "counts")

\S4method{simulateDE}{SingleCellExperiment}(x, which_cols, ...)
}
\arguments{
\item{x}{A numeric matrix containing features in rows and cells in columns.
Alternatively, a \linkS4class{SummarizedExperiment} or
\linkS4class{SingleCellExperiment} object.}

\item{which_cols}{Character, numeric or logical vector specifying for which
columns (cells, samples, ...) of \code{x} the feature swapping should occur.
Usually, this will be a random subset of columns belonging to a mock group.}

\item{...}{For the generic, arguments to be passed to specific methods.

For the \code{SummarizedExperiment} method, further arguments to be passed to
the \code{ANY} method.

For the \code{SingleCellExperiment} method, further arguments to be passed to
the \code{SummarizedExperiment} method.}

\item{prop_DE}{Numeric scalar specifying the proportion of features that will
be simulated as DE. Default: \code{0.01}.}

\item{use_assay}{A string or integer scalar specifying the assay of \code{x} to be
used as input for the simulation. The default is to use \code{"counts"}.}
}
\value{
A \linkS4class{SummarizedExperiment} object with DE induced between the
specified \code{groups}. The simulated counts are in the \code{"counts"} assay of the
returned object. The \code{"sim_group"} column of the \code{colData} indicates whether
swapping was performed in that column (specified by the \code{which_cols}
argument). The \code{rowData} contains the
following columns:
\itemize{
\item \code{"is_DE"}: logical vector indicating the ground truth status of each
feature
\item \code{"swapped_feature"}: character vector indicating the original feature with
which the feature was swapped
}

If \code{x} was a \emph{SummarizedExperiment} object, the original \code{colData} and
\code{rowData} are combined with these new columns.

If \code{x} was a \emph{SingleCellExperiment} object, the output will also be a
\emph{SingleCellExperiment}.
}
\description{
A simple DE simulator that takes a dataset with a grouping factor and induces
DE by swapping some features in one of the groups but not the other. This
should ensure keeping the same data structure as the original data without
having to estimate distribution parameters.
}
\details{
Note that for \emph{SummarizedExperiment} objects, any additional assays
(containing for example normalized values) are not retained. For simple
library size-factor normalization, size factors for each cell will be
identical before and after simulation. However, for more sophisticated
methods, such as normalization by deconvolution, the size factors might
differ due to the feature swapping.
The expected number of DE features will be equal to \code{prop_DE * nrow(x)}. Note
however that it's possible that the actual number might be 1 lower than this.
This can happen when a single feature remains un-"swapped" without any other
features left to swap it with.
}
\examples{
example_sce <- scuttle::mockSCE()

## Swap features in one of the "treat1" group
cells_to_swap <- example_sce$Treatment == "treat1"

sim <- simulateDE(example_sce, which_cols = cells_to_swap, prop_DE = 0.1)
sim

}
\author{
Milan Malfait
}
