% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapper-package.R
\docType{package}
\name{swapper-package}
\alias{swapper}
\alias{swapper-package}
\title{swapper: Simulate DE Signal By Feature Swapping}
\description{
Provides a simple simulation method for high-throughput data. The core method consists of randomly scrambling features within one experimental group. This approach retains the reference data characteristics without requiring extensive model fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\author{
\strong{Maintainer}: Milan Malfait \email{milan.malfait94@gmail.com} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})

}
\keyword{internal}
