% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDE.R
\name{simulateDE}
\alias{simulateDE}
\alias{simulateDE,ANY-method}
\alias{simulateDE,SummarizedExperiment-method}
\title{Simulate DE by swapping genes}
\usage{
simulateDE(x, groups, ...)

\S4method{simulateDE}{ANY}(x, groups, prop_DE = 0.01)

\S4method{simulateDE}{SummarizedExperiment}(x, groups, ..., use_assay = "counts")
}
\arguments{
\item{x}{A numeric matrix containing features in rows and cells in columns.
Alternatively, a \linkS4class{SummarizedExperiment} object.}

\item{groups}{A vector of length equal to \code{ncol(x)} specifying the group to
which each cell is assigned. DE will be induced between these groups.}

\item{...}{For the generic, arguments to be passed to specific methods.

For the \code{SummarizedExperiment} method, further arguments to be passed to
the \code{ANY} method.}

\item{prop_DE}{Numeric scalar specifying the proportion of genes that will be
simulated as DE. Default: \code{0.01}.}

\item{use_assay}{A string or integer scalar specifying the assay of \code{x} to be
used as input for the simulation. The default is to use \code{"counts"}.}
}
\value{
A \linkS4class{SummarizedExperiment} object with DE induced between the
specified \code{groups}. The simulated counts are in the \code{"counts"} assay of the
returned object. The \code{"groups"} used as input for the simulation are recorded
in the \code{"sim_group"} column of the \code{colData}. The \code{rowData} contains the
following columns:
\itemize{
\item \code{"is_DE"}: logical vector indicating the ground truth status of each gene
\item \code{"swapped_gene"}: character vector indicating the original gene with which
the gene was swapped
}

The expected number of DE genes will be equal to \code{prop_DE * nrow(x)}. Note
however that it's possible that the actual number might be 1 lower than this.
This can happen when a single feature remains un-"swapped" without any other
features left to swap it with.

If \code{x} was a \emph{SummarizedExperiment} object, the original \code{colData} and
\code{rowData} are combined with these new columns.
}
\description{
A simple DE simulator that takes a dataset with a grouping factor and induces
DE by swapping some genes in one of the groups but not the other. This should
ensure keeping the same data structure as the original data without having to
estimate distribution parameters.
}
\details{
Note that for \emph{SummarizedExperiment} objects, any additional assays
(containing for example normalized values) are not retained. For simple
library size-factor normalization, size factors for each cell will be
identical before and after simulation. However, for more sophisticated
methods, such as normalization by deconvolution, the size factors might
differ due to the gene swapping.
}
\examples{
example_sce <- scuttle::mockSCE()
sim <- simulateDE(example_sce, groups = example_sce$Treatment, prop_DE = 0.1)
sim

}
\author{
Milan Malfait
}
